Update Uno dependencies and project configurations
Updated the `Uno.WinUI` package to `6.4.0-dev.235` and the `Uno.Sdk` versions in `global.json` to `6.4.0-dev.50` and `6.4.0-dev.256`.

Changed the project SDK in `MonacoEditorTestApp.csproj` from `Uno.Sdk.Private` to `Uno.Sdk` and upgraded the target framework to `net10.0-browserwasm`.

Commented out unused embedded resources in `MonacoEditorComponent.csproj` and added the `<WasmFingerprintAssets>` property with a value of `false` to optimize asset handling.